knowledge_base/requirements.txt
## Optional: faiss-cpu requires native deps (often `swig`) on macOS.
## If you need FAISS retrieval, install it separately (e.g., via conda/brew + pip).
# faiss-cpu
numpy==1.26.3
# OpenAI SDK: use a modern version for gpt-5.x compatibility (max_completion_tokens, etc.)
openai>=1.50.0
scikit-learn
sentence-transformers==2.2.2
tenacity==8.2.3
tiktoken==0.5.1 
torch
transformers==4.48.0
umap-learn==0.5.5
urllib3>=2.6.3

# Multimodal ingestion dependencies
# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# File processing
pypdf>=3.17.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic>=0.4.27

# Media processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# APIs (optional but recommended)
anthropic>=0.18.0

# Utilities
langdetect>=1.0.9
python-dateutil>=2.8.2
pydantic>=2.5.0

# Database extractors (optional)
sqlalchemy>=2.0.0
pymongo>=4.6.0

# Rate limiting and async
aiohttp>=3.13.3
ratelimit>=2.2.1

# Security patches (pyasn1, jaraco.context, starlette, aiohttp) applied in Dockerfile

# Progress bars (used when RAPTOR_PROGRESS=1)
tqdm>=4.66.0

# API server (used by api_server.py and serve scripts)
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.26.0
